--- system/bare/sys_mem/SConscript
@@ -1,10 +1,0 @@
-from building import *
-import os
-
-cwd     = GetCurrentDir()
-src     = Glob('*.c')
-CPPPATH = [cwd + '/../include/']
-
-group = DefineGroup('__mlibc_sys_bare', src, depend = [''], CPPPATH = CPPPATH)
-
-Return('group')
