+++ system/rtthread/SConscript
@@ -1,13 +1,10 @@
-import os
 from building import *
+import os
 
-cwd = GetCurrentDir()
-objs = []
-list = os.listdir(cwd)
+cwd     = GetCurrentDir()
+src     = Glob('*.c')
+CPPPATH = []
 
-for d in list:
-    path = os.path.join(cwd, d)
-    if os.path.isfile(os.path.join(path, 'SConscript')):
-        objs = objs + SConscript(os.path.join(d, 'SConscript'))
+group = DefineGroup('mlibc', src, depend = [''], CPPPATH = CPPPATH)
 
-Return('objs')
+Return('group')
