--- system/rtthread/sys_fio/__mlibc_sys_read.c
@@ -1,34 +1,0 @@
-/*
- * Copyright (c) mlibc & plct lab
- *
- * SPDX-License-Identifier: MIT
- *
- * Change Logs:
- * Date           Author       Notes
- * 2024/5/6    0Bitbiscuits  realize sys_read
- */
-#include <libc_config.h>
-#ifdef MLIBC_RUNNING_RTTHREAD
-
-#include <rtconfig.h>
-#ifdef RT_USING_DFS
-
-#include <sys/sys_fio.h>
-#include <dfs_file.h>
-#include <unistd.h>
-
-/**
- * @brief The function is used to read data from a file descriptor into the buffer.
- * 
- * @param fd the file descriptor
- * @param buf buffer
- * @param buf_size buffer size
- * @return ssize_t The number of bytes read, which returns -1 if the read operation fails.
- */
-ssize_t __mlibc_sys_read(int fd, void* buf, size_t buf_size)
-{
-    return read(fd, buf, buf_size);
-}
-
-#endif /* RT_USING_DFS */
-#endif /* MLIBC_RUNNING_RT_THREAD */
